fix: 修复接收端 session_id 符号扩展导致卡在恢复中0%
根因: JNI/WASM 端 session_id 拼接时 i64→u128 符号扩展，
当 lo 部分最高位为1时 hi 被污染为全F，导致 ~50% 的文件
接收端永远 session mismatch。

修复:
- jni.rs: session_id_lo/hi 先转 u64 再转 u128
- wasm.rs: 同上修复
- progress.rs: 新增 session_mismatch_streak 字段
- receiver.rs: 追踪连续 session mismatch 计数
- ReceiverSessionManager.kt: 仅从 descriptor 帧初始化 receiver，
  且检测持续 mismatch 时自动重建
- options.tsx: 设置 document.title 修复标签页显示 undefined
- package.json: manifest 增加 action.default_title

+++ crates/transfer-engine/src/wasm.rs
@@ -50,7 +50,7 @@ impl SenderSessionWasm {
         compression: u8,
     ) -> Result<SenderSessionWasm, JsValue> {
         _start();
-        let sid = SessionId(((session_id_hi as u128) << 64) | session_id_lo as u128);
+        let sid = SessionId(((session_id_hi as u64 as u128) << 64) | session_id_lo as u64 as u128);
         let cfg = SenderConfig {
             codec: Config { symbol_size },
             redundancy_pct,
